team-server/changelog/requirements.txt
yamale==2.2.0
pyyaml==5.4.1
natsort==7.0.1
requests==2.23.0